backend/ExpenseTracker/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-filter==23.3
django-storages==1.14.2
djangorestframework==3.14.0
pillow==10.2.0
python-dateutil==2.8.2
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
